Make invalid Issue impossible
- Issue.create will verify validity of criteria/remedies prior to building an Issue.
