+++ Sources/iTunes/Repair/Track+Issue.swift
@@ -96,7 +96,6 @@ extension Track {
   }
 
   func repair(_ issue: Issue) -> Track? {
-    guard issue.isValid else { return self }
     guard criteriaApplies(issue.criteria) else { return self }
 
     var fixedTrack: Track? = self
